+++ src/dipdup/http.py
@@ -17,6 +17,7 @@ from fcache.cache import FileCache  # type: ignore
 
 from dipdup import __version__
 from dipdup.config import HTTPConfig  # type: ignore
+from dipdup.prometheus import Metrics
 
 safe_exceptions = (
     aiohttp.ClientConnectionError,
@@ -102,6 +103,7 @@ class _HTTPGateway:
             raise RuntimeError('aiohttp session is closed')
         return self.__session
 
+    # TODO: Move to separate method to cover SignalR negotiations too
     async def _retry_request(self, method: str, url: str, weight: int = 1, **kwargs):
         """Retry a request in case of failure sleeping according to config"""
         attempt = 1
@@ -123,13 +125,20 @@ class _HTTPGateway:
                     raise e
 
                 ratelimit_sleep: Optional[float] = None
-                if isinstance(e, aiohttp.ClientResponseError) and e.status == HTTPStatus.TOO_MANY_REQUESTS:
-                    # NOTE: Sleep at least 5 seconds on ratelimit
-                    ratelimit_sleep = 5
-                    # TODO: Parse Retry-After in UTC date format
-                    with suppress(KeyError, ValueError):
-                        e.headers = cast(Mapping, e.headers)
-                        ratelimit_sleep = int(e.headers['Retry-After'])
+                if isinstance(e, aiohttp.ClientResponseError):
+                    if Metrics.enabled:
+                        Metrics.set_http_error(self._url, e.status)
+
+                    if e.status == HTTPStatus.TOO_MANY_REQUESTS:
+                        # NOTE: Sleep at least 5 seconds on ratelimit
+                        ratelimit_sleep = 5
+                        # TODO: Parse Retry-After in UTC date format
+                        with suppress(KeyError, ValueError):
+                            e.headers = cast(Mapping, e.headers)
+                            ratelimit_sleep = int(e.headers['Retry-After'])
+                else:
+                    if Metrics.enabled:
+                        Metrics.set_http_error(self._url, 0)
 
                 self._logger.warning('HTTP request attempt %s/%s failed: %s', attempt, retry_count_str, e)
                 self._logger.info('Waiting %s seconds before retry', ratelimit_sleep or retry_sleep)