+++ src/dipdup/scheduler.py
@@ -1,4 +1,3 @@
-from contextlib import AsyncExitStack
 from datetime import datetime
 from functools import partial
 
@@ -15,7 +14,6 @@ from dipdup.context import DipDupContext
 from dipdup.context import HookContext
 from dipdup.exceptions import ConfigurationError
 from dipdup.utils import FormattedLogger
-from dipdup.utils.database import in_global_transaction
 
 jobstores = {
     'default': MemoryJobStore(),
@@ -46,14 +44,10 @@ def add_job(ctx: DipDupContext, scheduler: AsyncIOScheduler, job_config: JobConf
             hook_config=hook_config,
         )
 
-        async with AsyncExitStack() as stack:
-            if hook_config.atomic:
-                await stack.enter_async_context(in_global_transaction())
+        await job_ctx.fire_hook(hook_config.callback, *args, **kwargs)
 
-            await job_ctx.fire_hook(hook_config.callback, *args, **kwargs)
-
-            if job_config.daemon:
-                raise ConfigurationError('Daemon jobs are intended to run forever')
+        if job_config.daemon:
+            raise ConfigurationError('Daemon jobs are intended to run forever')
 
     logger = FormattedLogger(
         name=f'dipdup.hooks.{hook_config.callback}',