+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 Please use [this](https://docs.gitlab.com/ee/development/changelog.html) document as guidelines to keep a changelog.
 
+## [unreleased]
+
+### Added
+
+* prometheus: Added ability to expose Prometheus metrics.
+* config: Added top-level config section `prometheus`. 
+
+### Fixed
+
+* context: Fixed `HookConfig.atomic` flag, which was ignored in `fire_hook` method.
+
 ## 4.1.1 - 2022-01-25
 
 ### Fixed
@@ -60,6 +71,7 @@ Please use [this](https://docs.gitlab.com/ee/development/changelog.html) documen
 * ci: Detect CI environment and skip tests that fail in GitHub Actions.
 * ci: Execute tests in parallel with `pytest-xdist` when possible.
 * ci: More strict linting rules of `flake8`.
+>>>>>>> master
 
 ## 4.0.3 - 2022-01-09